requirements.txt
flask
flask-sqlalchemy
groq
python-dotenv
flask-cors
psycopg2-binary
flask-migrate
gunicorn
